back/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
cffi==1.17.1
cssselect2==0.8.0
fonttools==4.58.4
pycparser==2.22
pyphen==0.17.2
tinycss2==1.4.0
weasyprint==65.1
fastapi-mail==1.5.0
tenacity==9.1.2
openpyxl==3.1.5
